requirements.txt
aiosqlite==0.20.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncclick==8.1.8.0
asyncpg==0.30.0
bcrypt==4.2.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dictdiffer==0.9.0
distlib==0.3.8
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
filelock==3.16.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iso8601==2.1.0
Mako==1.3.8
MarkupSafe==3.0.2
platformdirs==4.3.2
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
pypika-tortoise==0.3.2
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
tomlkit==0.13.2
tortoise-orm==0.23.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.26.4
